+++ lib/liboqs.dart
@@ -3,6 +3,5 @@ library;
 export 'src/exception.dart';
 export 'src/kem.dart';
 export 'src/liboqs.dart';
-export 'src/platform/library_loader.dart';
 export 'src/random.dart';
 export 'src/signature.dart';