+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add support "@Native" annotation instead use library loader class
+
 ## [1.0.2] - 2025-01-14
 
 ### Added